+++ addon/components/flexberry-groupedit.js
@@ -709,7 +709,7 @@ export default FlexberryBaseComponent.extend({
       ...
       actions: {
         ...
-        confirmDeleteRow() {
+        confirmDeleteRow(data) {
           return confirm('You sure?');
         }
         ...
@@ -718,6 +718,7 @@ export default FlexberryBaseComponent.extend({
       ```
 
     @method confirmDeleteRow
+    @param {Object} data selected row data.
     @return {Boolean} If `true` then delete row, else cancel.
   */
   confirmDeleteRow: undefined,
@@ -740,7 +741,7 @@ export default FlexberryBaseComponent.extend({
       ...
       actions: {
         ...
-        confirmDeleteRows() {
+        confirmDeleteRows(data) {
           return confirm('You sure?');
         }
         ...
@@ -749,6 +750,7 @@ export default FlexberryBaseComponent.extend({
       ```
 
     @method confirmDeleteRows
+    @param {Object} data selected rows data.
     @return {Boolean} If `true` then delete selected rows, else cancel.
   */
   confirmDeleteRows: undefined,