+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+* Change confirm delete rows actions on async.
 
 ## [2.4.0] - 2020-05-18
 ### Added